Simplify use of --plugin argument of BackupDumper
Since I118b92dd16fff40b253c2c1009827086ef847dc2 the classes are part of
the autoloader, so there is no need for a php file to specifiy.

This also makes the test pass on windows.
The given $IP contains the driver letter followed by a colon
(C:\Users\...).
But the string for --plugin is exploded on the colon and than there are
too much array entries, failing the test.

Change-Id: Ie4b10659730bc4bc07e59eb57fb95ee05d572dba