Remove redundant Stub class not compatible with PHPUnit any more
I'm not sure why this was introduced. Using
willReturnOnConsecutiveCalls() works just fine. The order of the two
expected calls is not going to change. So yes, this test is now
asserting that the two calls are done in exactly this order, even if
this detail is entirely irrelevant for the functionality of the code.
However, I don't think it's worth having so much special case code
just to not have this extra assertion when it's not a problem.

It's broken anyway and does not work with the PHPUnit version we are
currently using.

Change-Id: I2aec1adb5ec38b662f6d4297d925be8408f8b435